progress and start working on tinydb repository

+++ libs/log/file_logger.py
@@ -1,3 +1,4 @@
+from pathlib import Path
 from libs.log.base_logger import ILogger, LogLevel
 from datetime import datetime
 
@@ -5,10 +6,14 @@ from datetime import datetime
 class FileLogger(ILogger):
     def __init__(self, name: str = "FileLogger", filepath: str = "log.txt"):
         super().__init__(name)
-        self.filepath = filepath
+        self.filepath = Path(filepath)
         self._create_file()
 
     def _create_file(self):
+        if self.filepath.exists():
+            self._log(LogLevel.INFO, f"Log file {self.filepath} already exists")
+            return
+
         with open(self.filepath, "w") as file:
             file.write("")
